+++ .gitignore
@@ -15,4 +15,3 @@ tog-ngac-crosscpp/*
 *RESULT*
 *Result*
 *result*
-pytest-coverage.txt
