+++ src/NGAC/ngac_types/user.py
@@ -91,3 +91,57 @@ class User(NgacObject):
 
     def __getitem__(self, index: int) -> UserAttribute:
         return self.attributes[index]
+
+
+def test_create_user():
+    """
+    Test user creation
+    """
+    user = User([], id="SomeUser")
+    assert user.id == "SomeUser"
+
+
+def test_create_with_attributes():
+    """
+    Test user creation with attributes
+    """
+    attributes = [
+        UserAttribute("value1"),
+        UserAttribute("value2"),
+        UserAttribute("value3"),
+    ]
+    user = User(attributes, id="SomeUser")
+    assert user.id == "SomeUser"
+    assert user.get_attributes() == attributes
+
+
+def test_list_like_user():
+    """
+    Test user class, list like operations
+    """
+    attributes = [
+        UserAttribute("value1"),
+        UserAttribute("value2"),
+        UserAttribute("value3"),
+    ]
+    user = User(attributes, id="SomeUser")
+
+    assert user[0] == attributes[0]
+    assert user[1] == attributes[1]
+
+    assert len(user) == 3
+    popped = user.pop(0)
+    assert popped == UserAttribute("value1")
+    assert len(user) == 2
+    user.push(UserAttribute("value4"))
+    assert len(user) == 3
+    assert user[2] == UserAttribute("value4")
+
+    for attribute in user:
+        assert attribute in attributes or attribute == UserAttribute("value4")
+
+
+if __name__ == "__main__":
+    test_create_user()
+    test_create_with_attributes()
+    test_list_like_user()
